Seperated playback and CSVReader
Made a seperate file for CSVReader
Playback calls the CSVReader and it returns datapoints for it to use
