+++ src/frontend/Assets/Scenes/Playback.unity
@@ -1240,7 +1240,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1259239079}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c033a158239f1474f9d1578e1b24fed0, type: 3}
+  m_Script: {fileID: 11500000, guid: cbf9ebbab98d03c4886b5a44b3c06192, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   playButton: {fileID: 920298906}
